ServiceStateManager corrections (#38)
