+++ lib/serviceStateManager/README.md
@@ -13,8 +13,9 @@ This module is responsible for the health check and graceful shutdown of dojot s
 
 # **Usage**
 
-Instantiate the `Manager` class and add shutdown handlers and health checkers to it. Whenever you
-want to gracefully shutdown the system, you call the `shutdown` function and you're done.
+Instantiate the `ServiceStateManager` class and add shutdown handlers and health checkers to it.
+Whenever you want to gracefully shutdown the system, you call the `shutdown` function and you're
+done. The shutdown function is also called whenever the service detects some signals, like `SIGTERM`.
 
 Beware that when you instantiate the Manager class, an HTTP server is created, by default, in the
 port `9000`. It is the health check server, created by the [lightship library](https://github.com/gajus/lightship/).
@@ -27,6 +28,11 @@ to use them.
 
 Don't forget to check the example in the [examples directory](../../examples/serviceStateManager)
 for a practical usage.
+
+## **Registering services**
+
+Before you add any health checker, you must register the services you will check. This is done by
+calling the `registerService` function.
 
 ## **Shutdown handlers**
 
@@ -45,19 +51,40 @@ can adapt it to your needs. The modes are:
 `signalReady` when the HTTP server receives the `listening` event and call `signalNotReady` when
 receiving `close` or `error` events)
 - Managed mode: using the module's internally managed health check interface via the
-`addHealthChecker` method.
+`addHealthChecker` method. The functions passed via this method should be asynchronous.
 
 It is recommended to create one health check per thing you want to check. For example: you have
 Kafka, VerneMQ and an HTTP server. Each one of these should have its own health checker (their names
 could be something like `kafka`, `verne` and `server`), since we don't want to block the Event Loop
 with long tasks. Each one can be configured with a different interval.
 
+## **Example**
+
+This is a quick example for consulting.For a better example, check the [examples directory](../../examples/serviceStateManager).
+
+```js
+const ServiceStateManager = require('@dojot/microservice-sdk');
+
+const manager = new ServiceStateManager();
+manager.registerService('mqtt');
+manager.registerService('http');
+
+// Remember that the health checkers should be asynchronous
+// You can add as many health checkers as you wish
+manager.addHealthChecker('mqtt', mqttHealthChecker, mqttInterval);
+manager.addHealthChecker('http', httpHealthChecker, httpInterval);
+
+// You can add as many shutdown handlers as you wish
+manager.registerShutdownHandler(mqttShutdown);
+manager.registerShutdownHandler(httpShutdown);
+```
+
 # **Configuration**
 
 The configuration is done via the `config` parameter when instantiating the `Manager` class.
 
-__NOTE THAT__ the logger used inside the Manager will indirectly inherit the configurations from the
-application logger, since the SDK Logger class is globally defined.
+__NOTE THAT__ the logger used inside the ServiceStateManager will indirectly inherit the
+configurations from the application logger, since the SDK Logger class is globally defined.
 
 ## **lightship**
 
@@ -70,6 +97,7 @@ The default configuration is:
 ```js
 {
   detectKubernetes: false,
+  terminate: () => { process.exit(0); },
 };
 ```
