--- lib/serviceStateManager/index.js
@@ -1,3 +1,0 @@
-const Manager = require('./Manager');
-
-module.exports = { Manager };
