+++ package.json
@@ -56,6 +56,7 @@
     "eslint": "^6.8.0",
     "eslint-config-airbnb-base": "^14.0.0",
     "eslint-plugin-import": "^2.19.1",
+    "@types/jest": "^26.0.15",
     "jest": "^26.6.3",
     "triple-beam": "^1.3.0"
   },
